requirements.txt
appnope==0.1.2
backcall==0.2.0
boto3==1.16.47
botocore==1.19.47
decorator==4.4.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
jmespath==0.10.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.7.0
Pygments==2.7.3
python-dateutil==2.8.1
python-dotenv==0.15.0
s3transfer==0.3.3
six==1.15.0
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
